requirements.txt
psycopg2-binary
Django==4.0.5
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
djangorestframework==3.13.1
dlib-binary==19.21.99
dnspython==2.2.1
numpy==1.23.0
Pillow==9.1.1
pymongo==4.1.1
pytz==2022.1
gunicorn
